Various changes
Added random data generator.
Fixed max number of pages on the pager to 5.
UI fixes (mainly against overflowing text).
More small changes.

+++ Teller.Models/UserInfo.cs
@@ -20,8 +20,9 @@
 
         public string AvatarPath { get; set; }
 
+        // was byte
         [Required]
-        public byte StoryViolations { get; set; }
+        public int StoryViolations { get; set; }
 
         [Required]
         public int CommentViolations { get; set; }
